move `_format_string()` verbatim from compiler.py to lex/parser.py
